requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
dateparser>=1.2.0
tkcalendar>=1.6.1